# docs/openapi.yaml
openapi: 3.0.3
info:
  title: vocagame technical testing backend
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /api/v1/auth/signup:
    post:
      summary: Register new user
      description: create new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 2
                  maxLength: 255
                email:
                  type: string
                  maxLength: 255
                password:
                  type: string
                  minLength: 7
                  maxLength: 50
              required:
                - username
                - email
                - password
            examples:
              success:
                description: Sign up user john doe
                value:
                  username: john doe
                  email: doe@mail.com
                  password: vocagame111
      responses:
        "201":
          description: Success register new user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithoutTokens"
              examples:
                success:
                  description: Success register new user
                  value:
                    description: Sign up success
                    error_message: ''
                    execute_at: 2024/09/22 19:45:05.469
                    result: success
                    value:
                      username: John Doe
                      email: doe@mail.com
        "409":
          description: Duplicate email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
              examples:
                duplicated email:
                  description: email is registered
                  value:
                    description: 
                    - this email address is already in use
                    error_message: Conflict
                    execute_at: 2024/11/05 16:29:11.639
                    result: failure
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
              examples:
                validationerror:
                  description: Validation error
                  value:
                    description:
                    - username is a required field
                    - Password must be at least 7 characters in length
                    error_message: Unprocessable Entity
                    execute_at: 2024/11/04 14:29:01.336
                    result: failure
  /api/v1/auth/login:
    post:
      summary: Login User
      description: Login user with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  maxLength: 255
                password:
                  type: string
                  minLength: 7
                  maxLength: 50
              required:
                - email
                - password
            examples:
              success:
                description: correct login parameter
                value:
                  email: doe@mail.com
                  password: johndoe
      responses:
        '200':
          description: Success login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithTokens"
                success:
                  description: Success login
                  value:
                    description: Login success
                    error_message: ''
                    execute_at: 2024/09/23 16:49:26.806
                    result: success
                    value:
                      username: john doe
                      email: doe@mail.com
                      access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                      refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
              examples:
                wrong password:
                  description: wrong password
                  value:
                    description: 
                    - password is wrong
                    error_message: Unauthorized
                    execute_at: 2024/09/22 19:44:11.778
                    result: failure
                wrong email:
                  description: wrong email
                  value:
                    description: 
                    - no user found with this email doee@mail.com
                    error_message: Unauthorized
                    execute_at: 2024/09/22 19:44:11.778
                    result: failure
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
              examples:
                validationerror:
                  description: Validation error
                  value:
                    description:
                    - Email must be a valid email address
                    error_message: Unprocessable Entity
                    execute_at: 2024/09/22 19:44:11.778
                    result: failure
  /api/v1/auth/logout:
    post:
      summary: Logout
      description: Logout user with auth header
      security:
      - bearerAuth: []
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
          format: bearer
        description: Bearer refresh_token
      responses:
        '200':
          description: success log out user
          content:
            text/plain:
              schema:
                type: boolean
              examples:
                success:
                  value: logout success
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
              examples:
                blacklist token:
                  value:
                    description: 
                    - token is blacklist
                    error_message: Unauthorized
                    execute_at: 2024/09/22 19:44:11.778
                    result: failure
                empty auth header:
                  description: empty auth header
                  value:
                    description:
                    - no authorization header found
                    error_message: Unprocessable Entity
                    execute_at: 2024/09/22 19:44:11.778
                    result: failure
                expired access token:
                  description: empty auth header
                  value:
                    description:
                    - 'failed to parse token when verifying, err: token has invalid claims: token is expired'
                    error_message: Unprocessable Entity
                    execute_at: 2024/09/22 19:44:11.778
                    result: failure
  /api/v1/auth/refresh_token:
    post:
      summary: Refresh token
      description: renew refresh and access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                access_token:
                  type: string
                refresh_token:
                  type: string
              required:
                - access_token
                - refresh_token
            example:
              access_token: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjJlMDcy...
              refresh_token: 4c489afa-485f-43f3-b5bc-23dd95b5a17a
      responses:
        '200':
          description: success renew refresh and access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseOnlyTokens"
              example:
                description: refresh token success
                error_message: ''
                execute_at: 2024/10/30 22:46:04.684
                result: success
                value:
                  refresh_token: 4d1aeb5b-24a9-4dc9-bb39-c796e4d73e78
                  access_token: Bearer eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjlkODhkM2...
        '401':
          description: invalid refresh token
          content:
            application/json:
              schema:
                type: object
                properties:
                  description:
                    type: string
                  error_message:
                    type: string
                  execute_at:
                    type: string
                  result:
                    type: string
              examples:
                Wrong Token:
                  description: access token & refresh token is different
                  value:
                    description:
                    - 'invalid refresh token, msg: failed to read refresh token, err:
                      no rows in result set'
                    error_message: Unauthorized
                    execute_at: 2024/10/30 22:47:38.331
                    result: failure
                Expire Token:
                  description: refresh token is expire
                  value:
                    description:
                    - refresh token is expire
                    error_message: Unauthorized
                    execute_at: 2024/10/30 22:47:38.331
                    result: failure
  /api/v1/auth/delete_user:
    delete:
      summary: Delete user
      description: Delete user from database using access token in auth header
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: Authorization
          schema:
            type: string
            format: bearer
          description: Bearer refresh_token
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      value:
                        type: string
              example:
                description: user deleted
                error_message: ""
                execute_at: 2024/09/22 19:44:11.778
                result: success
                value: ""
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
              examples:
                blacklist token:
                  value:
                    description: 
                    - token is blacklist
                    error_message: Unauthorized
                    execute_at: 2024/09/22 19:44:11.778
                    result: failure
                expired access token:
                  description: empty auth header
                  value:
                    description:
                    - 'failed to parse token when verifying, err: token has invalid claims: token is expired'
                    error_message: Unprocessable Entity
                    execute_at: 2024/09/22 19:44:11.778
                    result: failure
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
              example:
                description: 
                - 'failed to delete user, err: no user found'
                error_message: Unauthorized
                execute_at: 2024/09/22 19:44:11.778
                result: failure
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    error:
      type: object
      properties:
        description:
          type: array
        error_message:
          type: string
        execute_at:
          type: string
        result:
          type: string
    BaseResponse:
      type: object
      properties:
        description:
          type: string
        error_message:
          type: string
        execute_at:
          type: string
        result:
          type: string
    ValueWithTokens:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        access_token:
          type: string
        refresh_token:
          type: string
    ValueWithoutTokens:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
    ValueOnlyTokens:
      type: object
      properties:
        refresh_token:
          type: string
        token_token:
          type: string
...
    ResponseWithTokens:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            value:
              $ref: '#/components/schemas/ValueWithTokens'
    ResponseWithoutTokens:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            value:
              $ref: '#/components/schemas/ValueWithoutTokens'
    ResponseOnlyTokens:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            value:
              $ref: '#/components/schemas/ValueOnlyTokens'
...